fix: do not delete data unless told otherwise (see #232)

+++ burpui/misc/parser/burp1.py
@@ -373,7 +373,8 @@ class Parser(Doc):
 
                 self._refresh_cache()
 
-            shutil.rmtree(data)
+            if delete:
+                shutil.rmtree(data)
 
         except OSError as exp:
             res.append([NOTIF_ERROR, str(exp)])
